change task on server side
