+++ server/modules/encryption.js
@@ -1,6 +1,7 @@
 const bcrypt = require('bcryptjs');
 
-const SALT_WORK_FACTOR = 10; // This determines how secure the salt should be
+// This determines how secure the salt should be
+const SALT_WORK_FACTOR = 10; 
 
 const encryptPassword = (password) => {
   const salt = bcrypt.genSaltSync(SALT_WORK_FACTOR); // This generates a random salt
@@ -15,7 +16,6 @@ const comparePassword = (candidatePassword, storedPassword) => {
   The stored password has the original salt, so it will run the
   candidate password and salt through the same hashing process as before.
   If that result is the same as the stored password, then we have a match!
-  If this interests you, check out this video https://www.youtube.com/watch?v=8ZtInClXe1Q
   */
   return bcrypt.compareSync(candidatePassword, storedPassword);
 };
