+++ server/routes/admin.router.js
@@ -4,11 +4,11 @@ const router = express.Router();
 
 // get Client List
 router.get('/client-list', (req, res) => {
-    console.log('in admin GET clientList:')
+    // console.log('in admin GET clientList:')
     const queryText = `SELECT * FROM "users" ORDER BY "id" ASC;`;
     pool.query(queryText)
     .then( (result) => {
-        console.table(result.rows);
+        // console.table(result.rows);
         res.send(result.rows);
     })
     .catch( (error) => {
