+++ server/routes/alert.router.js
@@ -5,7 +5,7 @@ const router = express.Router();
 
 // GET Alert List
 router.get('/list/all', (req, res) => {
-    console.log('in alert GET all alerts:')
+    // console.log('in alert GET all alerts:')
     const queryText = `SELECT "alerts"."id", "description", "user_id", TO_CHAR("created", 'FMDay, FMMonth FMDD FMYYYY HH12:MI am') as created,
     "resolved", "users"."firstname" as firstname, "users"."lastname" as lastname, "teams"."name" as teamname, "teams"."id" as team_id
     FROM "alerts" 
@@ -14,7 +14,7 @@ router.get('/list/all', (req, res) => {
     ORDER BY "id" DESC;`;
     pool.query(queryText)
     .then( (result) => {
-        console.table(result.rows);
+        // console.table(result.rows);
         res.send(result.rows);
     })
     .catch( (error) => {
@@ -25,7 +25,7 @@ router.get('/list/all', (req, res) => {
 
 // SET alert resolved 
 router.put('/resolve/:id', (req, res) => {
-    console.log('in resolve alert PUT with id:', req.params.id);
+    // console.log('in resolve alert PUT with id:', req.params.id);
     const queryText = `UPDATE "alerts" SET "resolved"=true WHERE "id"=$1;`;
     pool.query(queryText, [Number(req.params.id)])
     .then( (result) => {
@@ -39,7 +39,7 @@ router.put('/resolve/:id', (req, res) => {
 
 // post new alert to db
 router.post('/new', (req, res) => {
-    console.log('in submit alert POST with', req.body);
+    // console.log('in submit alert POST with', req.body);
     const id = req.body.id;
     const desc = req.body.description;
     const queryText = `INSERT INTO "alerts" ("type", "description", "user_id") VALUES ('text', $1, $2)`;
